refactor: fix migrate on command

+++ app/Console/Commands/SeedPools.php
@@ -22,7 +22,9 @@ class SeedPools extends Command
 
         $this->info('Seeding pools...');
 
-        $this->callSilently('migrate:fresh', ['--force']);
+        if (! DB::transactionLevel()) {
+            $this->callSilently('migrate:fresh', ['--force']);
+        }
 
         DB::transaction(function (): void {
             $years = $this->scandir('pools');
